src/delaystore/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
blinker==1.9.0
    # via flask
cachetools==5.3.2
    # via google-auth
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.2.1
    # via flask
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
flask==3.1.1
    # via -r requirements.in
google-api-core==2.12.0
    # via -r requirements.in
google-auth==2.23.4
    # via google-api-core
googleapis-common-protos==1.61.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio==1.59.2
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-health-checking==1.59.2
    # via -r requirements.in
grpcio-tools==1.59.2
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
opentelemetry-api==1.20.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
opentelemetry-distro==0.41b0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.20.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.20.0
    # via -r requirements.in
opentelemetry-instrumentation==0.41b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-grpc==0.41b0
    # via -r requirements.in
opentelemetry-proto==1.20.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.20.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation-grpc
opentelemetry-semantic-conventions==0.41b0
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
protobuf==4.25.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   grpcio-tools
    #   opentelemetry-proto
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-json-logger==2.0.7
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
rsa==4.9
    # via
    #   -r requirements.in
    #   google-auth
typing-extensions==4.8.0
    # via opentelemetry-sdk
urllib3==2.0.7
    # via requests
werkzeug==3.1.3
    # via flask
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
